+++ x/programs/rust/wasmlanche-sdk/src/program.rs
@@ -5,39 +5,6 @@ use crate::{
 use borsh::{BorshDeserialize, BorshSerialize};
 use std::hash::Hash;
 
-#[macro_export]
-macro_rules! dbg {
-    () => {
-        if cfg!(debug_assertions) {
-            let as_string = format!("[{}:{}:{}]", file!(), line!(), column!());
-            $crate::log(as_string).unwrap();
-        }
-    };
-    ($val:expr $(,)?) => {{
-        match $val {
-            tmp => {
-                #[cfg(debug_assertions)]
-                {
-                    let as_string = format!(
-                        "[{}:{}:{}] {} = {:#?}",
-                        file!(),
-                        line!(),
-                        column!(),
-                        stringify!($val),
-                        &tmp
-                    );
-                    $crate::log(as_string).unwrap();
-                }
-                tmp
-            }
-        }
-    }};
-    ($($val:expr),+ $(,)?) => {
-        #[cfg(debug_assertions)]
-        ($($crate::dbg!($val)),+,)
-    };
-}
-
 /// Represents the current Program in the context of the caller. Or an external
 /// program that is being invoked.
 #[derive(Clone, Copy, BorshDeserialize, BorshSerialize, Debug)]
